Wheel spin
Wheel spin implemented. The wheel has a spin animation, where max speed and duration can be controlled. It also calculates a winner, and asks whether the winner should be removed from the list or not.

An arrow needs to be added above the wheel, and the calculation for the winner needs to be changed as it is currently inconsistent.