Fixed winner selection
There was a bug causing the incorrect winner to be displayed when the wheel stops spinning. This was due to the items being indexed in the reverse order as they appear on-screen, alongside the first item appearing to the right of the wheel, not at the top of the wheel as expected.